test(hooks): keep delivery coverage beside its owner
Move user-input delivery coverage out of the oversized lifecycle test module
so the inline-test invariant remains enforced.
